requirements.txt
anyio==4.11.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.10.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
python-telegram-bot==13.15
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.15